# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stream Lights personal lighting control API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Device {
  id: number;
  name: string;
  /** philips_hue | lifx | govee | nanoleaf | wiz | kasa | generic_http */
  type: string;
  /** @nullable */
  bridgeIp?: string | null;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  deviceId?: string | null;
  enabled: boolean;
  isOnline: boolean;
  /** @nullable */
  currentColor?: string | null;
  /** @nullable */
  brightness?: number | null;
  createdAt?: string;
}

export interface DeviceInput {
  name: string;
  type: string;
  bridgeIp?: string;
  apiKey?: string;
  deviceId?: string;
  enabled?: boolean;
}

export interface DeviceUpdate {
  name?: string;
  bridgeIp?: string;
  apiKey?: string;
  deviceId?: string;
  enabled?: boolean;
}

export interface Platform {
  id: number;
  /** twitch | youtube | kick | streamlabs | streamelements */
  platform: string;
  /** @nullable */
  channelName?: string | null;
  connected: boolean;
  /** @nullable */
  clientId?: string | null;
  eventsEnabled?: boolean;
  createdAt?: string;
}

export interface PlatformConnect {
  channelName: string;
  clientId?: string;
  clientSecret?: string;
  accessToken?: string;
}

export interface Trigger {
  id: number;
  name: string;
  /** follow | subscribe | bits | donation | raid | host | ban | timeout | chat_message | channel_point */
  eventType: string;
  /** @nullable */
  platform?: string | null;
  enabled: boolean;
  /** Hex color like */
  color?: string;
  brightness?: number;
  durationMs?: number;
  /** solid | strobe | pulse | rainbow | fade | police */
  effect?: string;
  returnToIdle?: boolean;
  /**
     * Minimum bits/donation to trigger
     * @nullable
     */
  minAmount?: number | null;
  deviceIds?: number[];
  createdAt?: string;
}

export interface TriggerInput {
  name: string;
  eventType: string;
  platform?: string;
  enabled?: boolean;
  color: string;
  brightness: number;
  durationMs: number;
  effect: string;
  returnToIdle?: boolean;
  minAmount?: number;
  deviceIds?: number[];
}

export interface TriggerUpdate {
  name?: string;
  platform?: string;
  enabled?: boolean;
  color?: string;
  brightness?: number;
  durationMs?: number;
  effect?: string;
  returnToIdle?: boolean;
  minAmount?: number;
  deviceIds?: number[];
}

export interface Command {
  id: number;
  name: string;
  /** !red, !party, etc. */
  command: string;
  color?: string;
  brightness?: number;
  durationMs?: number;
  effect?: string;
  enabled: boolean;
  cooldownSeconds?: number;
  usageCount?: number;
  createdAt?: string;
}

export interface CommandInput {
  name: string;
  command: string;
  color: string;
  brightness: number;
  durationMs: number;
  effect: string;
  enabled?: boolean;
  cooldownSeconds?: number;
}

export interface CommandUpdate {
  name?: string;
  command?: string;
  color?: string;
  brightness?: number;
  durationMs?: number;
  effect?: string;
  enabled?: boolean;
  cooldownSeconds?: number;
}

export interface ActivityEntry {
  id: number;
  eventType: string;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  colorTriggered?: string | null;
  /** @nullable */
  effectTriggered?: string | null;
  triggeredAt: string;
}

export interface LightEffect {
  color: string;
  brightness: number;
  durationMs: number;
  effect: string;
  deviceIds?: number[];
}

export interface ActionResult {
  success: boolean;
  message: string;
}

export interface EventTypeCount {
  eventType: string;
  count: number;
}

export interface DashboardStats {
  totalDevices: number;
  onlineDevices: number;
  activePlatforms: number;
  totalTriggers: number;
  activeTriggers: number;
  totalCommands: number;
  eventsToday: number;
  topEventTypes: EventTypeCount[];
}

export interface Settings {
  id: number;
  globalEnabled: boolean;
  idleColor: string;
  idleBrightness: number;
  idleEnabled: boolean;
  transitionSpeed?: number;
  notificationsEnabled?: boolean;
  overlayEnabled?: boolean;
  overlayPort?: number;
}

export interface SettingsUpdate {
  globalEnabled?: boolean;
  idleColor?: string;
  idleBrightness?: number;
  idleEnabled?: boolean;
  transitionSpeed?: number;
  notificationsEnabled?: boolean;
  overlayEnabled?: boolean;
  overlayPort?: number;
}

export type ListActivityParams = {
limit?: number;
eventType?: string;
};