# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stream Lights personal lighting control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all smart light devices
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Add a new smart light device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get a device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    patch:
      operationId: updateDevice
      tags: [devices]
      summary: Update a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /devices/{id}/test:
    post:
      operationId: testDevice
      tags: [devices]
      summary: Send a test flash to a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LightEffect"
      responses:
        "200":
          description: Test triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /devices/{id}/toggle:
    post:
      operationId: toggleDevice
      tags: [devices]
      summary: Toggle device on/off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
...
  /platforms/{platform}/connect:
    post:
      operationId: connectPlatform
      tags: [platforms]
      summary: Connect a streaming platform
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformConnect"
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
  /platforms/{platform}/disconnect:
    post:
      operationId: disconnectPlatform
      tags: [platforms]
      summary: Disconnect a streaming platform
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
...
  /commands:
    get:
      operationId: listCommands
      tags: [commands]
      summary: List all chat commands
      responses:
        "200":
          description: Commands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"
    post:
      operationId: createCommand
      tags: [commands]
      summary: Create a chat command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
  /commands/{id}:
    patch:
      operationId: updateCommand
      tags: [commands]
      summary: Update a chat command
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
    delete:
      operationId: deleteCommand
      tags: [commands]
      summary: Delete a chat command
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Device:
      type: object
      required: [id, name, type, enabled, isOnline]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "philips_hue | lifx | govee | nanoleaf | wiz | kasa | generic_http"
        bridgeIp:
          type: ["string", "null"]
        apiKey:
          type: ["string", "null"]
        deviceId:
          type: ["string", "null"]
        enabled:
          type: boolean
        isOnline:
          type: boolean
        currentColor:
          type: ["string", "null"]
        brightness:
          type: ["integer", "null"]
        createdAt:
          type: string
    DeviceInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        bridgeIp:
          type: string
        apiKey:
          type: string
        deviceId:
          type: string
        enabled:
          type: boolean
    DeviceUpdate:
      type: object
      properties:
        name:
          type: string
        bridgeIp:
          type: string
        apiKey:
          type: string
        deviceId:
          type: string
        enabled:
          type: boolean
    Platform:
      type: object
      required: [id, platform, connected]
      properties:
        id:
          type: integer
        platform:
          type: string
          description: "twitch | youtube | kick | streamlabs | streamelements"
        channelName:
          type: ["string", "null"]
        connected:
          type: boolean
        clientId:
          type: ["string", "null"]
        eventsEnabled:
          type: boolean
        createdAt:
          type: string
    PlatformConnect:
      type: object
      required: [channelName]
      properties:
        channelName:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        accessToken:
          type: string
...
    Command:
      type: object
      required: [id, command, name, enabled]
      properties:
        id:
          type: integer
        name:
          type: string
        command:
          type: string
          description: "!red, !party, etc."
        color:
          type: string
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
        enabled:
          type: boolean
        cooldownSeconds:
          type: integer
        usageCount:
          type: integer
        createdAt:
          type: string
    CommandInput:
      type: object
      required: [name, command, color, brightness, durationMs, effect]
      properties:
        name:
          type: string
        command:
          type: string
        color:
          type: string
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
        enabled:
          type: boolean
        cooldownSeconds:
          type: integer
    CommandUpdate:
      type: object
      properties:
        name:
          type: string
        command:
          type: string
        color:
          type: string
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
        enabled:
          type: boolean
        cooldownSeconds:
          type: integer
...
    LightEffect:
      type: object
      required: [color, brightness, durationMs, effect]
      properties:
        color:
          type: string
        brightness:
          type: integer
        durationMs:
          type: integer
        effect:
          type: string
        deviceIds:
          type: array
          items:
            type: integer
    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...