# lib/api-zod/src/generated/types/commandUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stream Lights personal lighting control API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandUpdate {
  name?: string;
  command?: string;
  color?: string;
  brightness?: number;
  durationMs?: number;
  effect?: string;
  enabled?: boolean;
  cooldownSeconds?: number;
}