# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stream Lights personal lighting control API
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceInput {
  name: string;
  type: string;
  bridgeIp?: string;
  apiKey?: string;
  deviceId?: string;
  enabled?: boolean;
}