# lib/api-zod/src/generated/types/platformConnect.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stream Lights personal lighting control API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformConnect {
  channelName: string;
  clientId?: string;
  clientSecret?: string;
  accessToken?: string;
}